Revert sum fix due to UFLD model failures, consolidate unit tests
